Extract useInterval
