+++ app/utils/useInterval.js
@@ -1,0 +1,18 @@
+import { useEffect, useRef } from 'react';
+
+export default function useInterval(callback, delay) {
+  const savedCallback = useRef();
+
+  useEffect(() => {
+    savedCallback.current = callback;
+  });
+
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+
+    let id = setInterval(tick, delay); // eslint-disable-line
+    return () => clearInterval(id);
+  }, [delay]);
+}
